不備改修24項目の並列実装（配線/UI/戦闘/テスト/パフォーマンス）

+++ Assets/MyAsset/Core/Common/Structs.cs
@@ -213,7 +213,7 @@ namespace Game.Core
         [FoldoutGroup("コスト")]
         [MinValue(0)] public float staminaCost;
         [FoldoutGroup("コスト")]
-        [MinValue(0)] public float mpCost;
+        [MinValue(0)] public int mpCost;
 
         [FoldoutGroup("ヒット")]
         [MinValue(1)] public int maxHitCount;
